api/requirements.txt
# Google requirements
google
google-cloud-vision
google-cloud-storage

# Flask extensions
flask
flask-cors
flask-security
flask-api

# Image Manipulation
opencv-python
imutils==0.4.6
pyzbar

# ORM/Database
sqlalchemy
mysqlclient
Flask-SQLAlchemy
Flask-Migrate
bcrypt


# MISC
requests
pyjwt
